third_party/WebKit/LayoutTests/http/tests/devtools/layers/layers-panel-mouse-events-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests moust hover/select events handling in the Layers panel

Hovering b1 in tree
Layer b1 in tree: hovered
Layer b3 in tree: 
Hovering b3 in tree
Layer b1 in tree: 
Layer b3 in tree: hovered
Hovering away from tree
Layer b1 in tree: 
Layer b3 in tree: 
Selecting b1 in tree
Layer b1 in tree: selected
Layer b3 in tree: 
Selecting b3 in tree
Layer b1 in tree: 
Layer b3 in tree: selected
